refactor: rename module following repo naming
