+++ ampd/src/handlers/mod.rs
@@ -6,7 +6,7 @@ pub mod evm_verify_msg;
 pub mod evm_verify_verifier_set;
 pub mod multisig;
 pub mod solana_verify_msg;
-pub mod solana_verify_worker_set;
+pub mod solana_verify_verifier_set;
 pub mod sui_verify_msg;
 pub mod sui_verify_verifier_set;
